Updated flags
DockEnabled_DonationsDock - Rearranged value overwrites
DockEnabled_MineDock - It has been discovered that the game's character management system breaks this dock. As a result the dock has been taken out of service until V6.1.0 which is expected to be update that fixes it.
DockEnabled_NoStandingJokeDock - Same as DockEnabled_MineDock.
Game_Version - Simply added an overwrite to it.
DayAndNight_TimeLockedOn - Simply removed an overwrite from it.